% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCellGraph.R
\name{BuildCellGraph}
\alias{BuildCellGraph}
\title{BuildCellGraph}
\usage{
BuildCellGraph(...)
}
\arguments{
\item{seurat_obj}{A Seurat object containing dimensionality reduction results.}

\item{reduction}{A string specifying which reduction to use (default is "pca").}
}
\value{
A cell-cell distance matrix.
}
\description{
Build cell–cell distance matrix from PCA embeddings
}
\details{
Build cell–cell distance matrix from PCA embeddings
This function extracts PCA embeddings from a Seurat object
and computes the Euclidean distance between all cells to
generate a distance matrix.
}
\examples{
\dontrun{
dists <- BuildCellGraph(seurat_obj, reduction = "pca")
}
}
