% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyCoreCells.R
\name{IdentifyCoreCells}
\alias{IdentifyCoreCells}
\title{Identify Core Cells}
\arguments{
\item{scRNA}{A Seurat object.}

\item{dist}{A distance matrix (e.g., Euclidean distance between cells).}

\item{top_n}{Number of nearest neighbors to consider (default = 20).}

\item{resolutions}{A numeric vector of resolution values used in clustering (e.g., c(0.1, 0.2, ..., 1.5)).}
}
\value{
A Seurat object with updated meta.data including CellClass annotations.
}
\description{
This function identifies "core" cells based on their local neighborhood consistency
across multiple clustering resolutions.
}
\details{
Identify Core Cells Based on Cluster Consistency
}
\examples{
\dontrun{
scRNA <- IdentifyCoreCells(scRNA, dist, top_n = 20, resolutions = seq(0.1, 1.5, 0.1))
}
}
